chore: 후기 영역 컴포넌트 분리 #130
