+++ src/components/feature/myProfile/MyprofileMain.tsx
@@ -1,10 +1,10 @@
 import { Tab, TabGroup, TabList, TabPanel, TabPanels } from '@headlessui/react';
 
 import MyKeyboardList from '@/components/feature/myProfile/MyKeyboardList';
-import MyReviewList from '@/components/feature/myProfile/MyReviewList';
+import MyReviewArea from '@/components/feature/myProfile/MyReviewArea';
 
 const MYPAGE_MENUS = [
-  { title: '내가 쓴 후기', component: MyReviewList },
+  { title: '내가 쓴 후기', component: MyReviewArea },
   { title: '내가 등록한 키보드', component: MyKeyboardList },
 ];
